django-project/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
# psycopg2-binary==2.9.9
python-dotenv==1.0.0
# supabase==2.3.0
python-telegram-bot==22.5
